Add payments

+++ DataAccessLibrary/DAO.cs
@@ -409,7 +409,7 @@ namespace DataAccessLibrary
         /// <param name="customerID"></param>
         /// <param name="amount"></param>
         /// <param name="treatmentCompleteDate"></param>
-        public static void AddNewpayment(int treatmentPlanID, string customerID, float amount, string treatmentCompleteDate)
+        public static void AddNewpayment(int treatmentPlanID, string customerID, decimal amount, string treatmentCompleteDate)
         {
             using (SqliteConnection db =
                 new SqliteConnection("Filename=dentalManagerDB.db"))
